+++ src/OpenTelemetry.Instrumentation.Runtime/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Remove Process related metrics from .NET Runtime metrics
+  ([#446](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/446))
+
 ## 0.2.0-alpha.1
 
 * Updated OTel SDK package version to 1.3.0