+++ src/OpenTelemetry.Instrumentation.Runtime/CHANGELOG.md
@@ -5,6 +5,10 @@
 * Update a few metrics to be in sync of doc updates for `gc.heap`,
   `gc.fragmentation.ratio`, `time.in.jit`, `process.cpu.count` and `assembly.count`
   ([#430](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/430))
+* Remove Process related metrics from .NET Runtime metrics
+  ([#446](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/446))
+* Add `exception.count` in Runtime metrics
+  ([#431](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/431))
 
 ## 0.2.0-alpha.1