+++ src/OpenTelemetry.Instrumentation.Runtime/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 0.3.0-alpha.1
+
 * Update a few metrics to be in sync of doc updates for `gc.heap`,
   `gc.fragmentation.ratio`, `time.in.jit`, `process.cpu.count` and `assembly.count`
   ([#430](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/430))
@@ -9,6 +11,8 @@
   ([#446](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/446))
 * Add `exception.count` in Runtime metrics
   ([#431](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/431))
+* Finalize GC instruments
+  ([#454](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/454))
 
 ## 0.2.0-alpha.1